feat: we have the "first" version of building
The basic goal here is to provide a build system that makes sense of the
pages directory.  this is a bit hard since that means we have to manage
the creation of mod files.

there are a ton of things that are not done, but i feel that this is the
first step in the right direction for alpha 1